Use same form for ScenarioNarrative
